libcoap-2.{map,sym}: create file in the top_builddir
If you are working out of tree you get a deadloop because you are unable
to recreate the *.{map,sym} file in your current working directory.

configure is thinking the source is already configured in another folder
if you want to call the configure script outside the top_sourcedir and
is requireing a "make distclean' call in top_sourcedir which is also
deleting libcoap-2.{map,sym}. But the source is *always* building these
two files in the top_sourcedir.

This commit is pointing the build of these two files into top_builddir
which is also top_sourcedir if you are doing a in tree build.